added isset checks to api

+++ api/deleteProduct.php
@@ -1,5 +1,11 @@
 <?php
 
+if (!isset($_POST["productID"])) {
+    http_response_code(400);
+    echo "Product ID not specified";
+    exit();
+}
+
 $productID = $_POST['productID'];
 
 $message = DeleteProduct($productID);
